Q3-followup-1: enable -Wconversion across the library
Re-enable -Wconversion in the strict warning set. Fix the conversion
sites in the library proper:

- src/util.h: now() casts long timeval fields to double explicitly.
- src/trie.h: int<->float conversions cast to float at the boundary.
- src/cinstancep2d.cc: introduce a static frand() helper that wraps
  urand() and returns float; use it everywhere a vec2/float receives
  a urand result.
- src/clinesp2d.cc, src/cliness2d.cc, src/crastss2d.cc, src/crastrs2d.cc,
  src/crasts2d.cc: replace double literals (0.0, 1.0) with their float
  equivalents inside float-valued expressions; drop the local `double
  factor` in LineRegion::set in favor of a float `factor`.

Doctest test files: switched M_PI references to the kPi/kHalfPi/kTwoPi
constants and added the util.h include where needed. tests/rast-test.cc
got partial fixes (literal suffixes, integer index casts) but still
has many double-literal call sites; exempted from -Werror via a
per-file Makefile rule.

Two files exempted via per-file -Wno-error rules in the Makefile:
- tests/rast-test.cc: regression-test driver, would need a full rewrite
  of urand()-heavy boilerplate.
- src/rast.cc: CLI driver with ~30 fgetenv("name", double_literal) call
  sites; slated for proper CLI11 conversion in a follow-up.

cedges.cc retains its existing -Wno-error exemption for the same
underlying reasons.

All three configs (debug / production / ultra) build cleanly. The
production and ultra doctest runs both pass 56/56 cases.

+++ src/trie.h
@@ -19,7 +19,7 @@ struct Trie1 {
 
   void init(float eps_, int w) {
     this->eps = eps_;
-    buckets.assign(int(w / eps_) + 1, {});
+    buckets.assign(int(static_cast<float>(w) / eps_) + 1, {});
   }
   void add(float x, T key) {
     int i = int(x / eps);
@@ -54,22 +54,22 @@ struct Trie2 {
     this->xoffset = xmin;
     this->yoffset = ymin;
     this->eps = eps_;
-    this->dimX = int(w / eps_) + 1;
-    this->dimY = int(h / eps_) + 1;
+    this->dimX = int(static_cast<float>(w) / eps_) + 1;
+    this->dimY = int(static_cast<float>(h) / eps_) + 1;
     buckets.assign(dimX * dimY, {});
   }
   void add(float x, float y, T key) {
-    x -= xoffset;
-    y -= yoffset;
+    x -= static_cast<float>(xoffset);
+    y -= static_cast<float>(yoffset);
     int i = int(x / eps);
     int j = int(y / eps);
     buckets[i * dimY + j].push_back({key, x, y});
   }
   void query(std::vector<T> &keys, float x0, float y0, float x1, float y1) {
-    x0 -= xoffset;
-    y0 -= yoffset;
-    x1 -= xoffset;
-    y1 -= yoffset;
+    x0 -= static_cast<float>(xoffset);
+    y0 -= static_cast<float>(yoffset);
+    x1 -= static_cast<float>(xoffset);
+    y1 -= static_cast<float>(yoffset);
     const int iEnd = clamp(int(x1 / eps), dimX);
     const int jEnd = clamp(int(y1 / eps), dimY);
     for (int i = clamp(int(x0 / eps), dimX); i <= iEnd; i++) {
