+++ src/util.h
@@ -115,7 +115,7 @@ inline int mkseed() {
 inline double now() {
   struct timeval tv;
   gettimeofday(&tv, nullptr);
-  return tv.tv_sec + 1e-6 * tv.tv_usec;
+  return static_cast<double>(tv.tv_sec) + 1e-6 * static_cast<double>(tv.tv_usec);
 }
 
 #endif  // RAST_SRC_UTIL_H_
