+++ tests/rast-test.cc
@@ -71,7 +71,7 @@ struct Segment {
 bool within(float x, float y, float eps) { return std::fabs(x - y) <= eps; }
 
 float uniform(float low, float high) { return float(drand48()) * (high - low) + low; }
-int iuniform(int low, int high) { return lrand48() % (high - low) + low; }
+int iuniform(int low, int high) { return static_cast<int>(lrand48() % (high - low)) + low; }
 vec2 vuniform(float x0, float y0, float x1, float y1) {
   return vec2(uniform(x0, x1), uniform(y0, y1));
 }
@@ -85,14 +85,14 @@ void test_linesp2d_1() {
       v = vuniform(0, 0, 512, 512);
     } while (distance(u, v) < 1.0f);
     Segment seg(u, v);
-    float tol = 0.1, atol = 1e-4;
+    float tol = 0.1f, atol = 1e-4f;
     lines->set_tolerance(tol, atol);
     float offset = seg.offset();
     float angle = seg.angle();
     int npoints = iuniform(2, 200);
     for (int i = 0; i < npoints; i++) {
-      vec2 p = seg.sample(i / float(npoints));
-      lines->add_ipoint(p[0], p[1], angle, 1.0);
+      vec2 p = seg.sample(static_cast<float>(i) / static_cast<float>(npoints));
+      lines->add_ipoint(p[0], p[1], angle, 1.0f);
     }
     lines->compute();
     if (lines->nresults() < 1) throw "didn't get any results";
@@ -119,15 +119,15 @@ void test_liness2d_1() {
       v = vuniform(0, 0, 512, 512);
     } while (distance(u, v) < 1.0f);
     Segment seg(u, v);
-    float tol = 0.1, atol = 1e-4;
+    float tol = 0.1f, atol = 1e-4f;
     lines->set_tolerance(tol, atol);
     float offset = seg.offset();
     float angle = seg.angle();
     int npoints = iuniform(2, 200);
     for (int i = 0; i < npoints; i++) {
-      vec2 p = seg.sample(i / float(npoints));
-      vec2 q = seg.sample(0.5f + i / float(npoints));
-      lines->add_iseg(p[0], p[1], q[0], q[1], angle, 1.0);
+      vec2 p = seg.sample(static_cast<float>(i) / static_cast<float>(npoints));
+      vec2 q = seg.sample(0.5f + static_cast<float>(i) / static_cast<float>(npoints));
+      lines->add_iseg(p[0], p[1], q[0], q[1], angle, 1.0f);
     }
     lines->compute();
     if (lines->nresults() < 1) throw "didn't get any results";
@@ -155,14 +155,14 @@ void test_rastp2d_1() {
     instance->set_nmodel_unoccluded(20);
     instance->set_error(0.0);
     instance->set_aerror(0.0);
-    instance->set_srange(0.5, 2.0);
+    instance->set_srange(0.5, 2.0f);
     instance->generate();
-    float tol = 1e-2;
-    float eps = 1.0;
-    float aeps = 0.1;
+    float tol = 1e-2f;
+    float eps = 1.0f;
+    float aeps = 0.1f;
     rast->set_min_q(0);
     rast->set_tolerance(tol);
-    rast->set_srange(0.5, 2.0);
+    rast->set_srange(0.5, 2.0f);
     assert(instance->nimage() == instance->nmodel());
     for (int i = 0; i < instance->nimage(); i++) {
       float x, y, a;
@@ -177,10 +177,10 @@ void test_rastp2d_1() {
     rast->match();
     assert(rast->nresults() > 0);
     assert(rast->ubound(0) == instance->nmodel());
-    assert(within(rast->translation(0, 0), instance->get_param(0), 2.0));
-    assert(within(rast->translation(0, 1), instance->get_param(1), 2.0));
-    assert(within(rast->angle(0), instance->get_param(2), 0.07));
-    assert(within(rast->scale(0), instance->get_param(3), 0.05));
+    assert(within(rast->translation(0, 0), instance->get_param(0), 2.0f));
+    assert(within(rast->translation(0, 1), instance->get_param(1), 2.0f));
+    assert(within(rast->angle(0), instance->get_param(2), 0.07f));
+    assert(within(rast->scale(0), instance->get_param(3), 0.05f));
   }
   end_trials;
 }
@@ -195,14 +195,14 @@ void test_rastp2d_2() {
     instance->set_nmodel_unoccluded(20);
     instance->set_error(0.0);
     instance->set_aerror(0.0);
-    instance->set_srange(0.5, 2.0);
+    instance->set_srange(0.5, 2.0f);
     instance->generate();
-    float tol = 1e-2;
-    float eps = 1.0;
-    float aeps = 0.1;
+    float tol = 1e-2f;
+    float eps = 1.0f;
+    float aeps = 0.1f;
     rast->set_min_q(0);
     rast->set_tolerance(tol);
-    rast->set_srange(0.5, 2.0);
+    rast->set_srange(0.5, 2.0f);
     rast->set_lsq(true);
     assert(instance->nimage() == instance->nmodel());
     for (int i = 0; i < instance->nimage(); i++) {
@@ -217,10 +217,10 @@ void test_rastp2d_2() {
     }
     rast->match();
     assert(rast->nresults() > 0);
-    assert(within(rast->translation(0, 0), instance->get_param(0), 0.1));
-    assert(within(rast->translation(0, 1), instance->get_param(1), 0.1));
-    assert(within(rast->angle(0), instance->get_param(2), 0.05));
-    assert(within(rast->scale(0), instance->get_param(3), 0.05));
+    assert(within(rast->translation(0, 0), instance->get_param(0), 0.1f));
+    assert(within(rast->translation(0, 1), instance->get_param(1), 0.1f));
+    assert(within(rast->angle(0), instance->get_param(2), 0.05f));
+    assert(within(rast->scale(0), instance->get_param(3), 0.05f));
   }
   end_trials;
 }
